update tag to   s.version          = '100.6562.2946' 1. order place api call firebase logs removed. 2. Available quanity. 3 firebase events logger removed 4, general api error added in elgrocer shopper

+++ el-grocer-shopper-sdk-iOS/Classes/Sources/Model/Entities+CRUD/ShoppingBasketItem+CRUD.swift
@@ -487,7 +487,7 @@ extension ShoppingBasketItem {
             return
         }
         if quantity > 0 {
-            guard product.availableQuantity != 0 else {
+            guard product.availableQuantity.intValue != 0 else {
                 NotificationCenter.default.post(name: KProductNotification, object: product)
                 return
             }
